chore(code): complete TODOs and fix minor issues
- Addressed multiple TODO
- Fixed minor environment error

+++ server/src/application/use_cases/tasks/DeleteTask.js
@@ -1,9 +1,13 @@
+const {PermissionDeniedError} = require("../../../errors");
+
 class DeleteTask {
     constructor(taskRepository) {
         this.taskRepository = taskRepository;
     }
 
-    async execute( {taskId} ) {
+    async execute( {taskId, ownerId} ) {
+        let task = this.taskRepository.getTaskById(taskId);
+        if (task.ownerId !== ownerId) throw new PermissionDeniedError();
         return await this.taskRepository.deleteTaskById(taskId);
     }
 }
